refactor(Toolbar): For now, I am disabling speech sythesis. More in the details...
I'm running into issues where when you change the page, it can mess up the play/pause function. You
need to click the restart button and play, pause to get it to play when going between pages. I don't
think it will be that tough to tackle, but at the moment I don't know if it's important to address.
The code still exists, its just conditionally not shown. Search for 5 == 0 to find where the code
is.